+++ src/components/AboutMeAll/AboutMeSamTas/animatedCard.tsx
