+++ src/components/AboutMeAll/AboutMeSamTas/styles/index.css
