+++ src/components/ui/Carousel/Card/styles/index.css
