+++ src/components/ui/Carousel/carouselHelper.ts
