+++ src/components/ui/TechStackDialog/styles/index.css
@@ -50,7 +50,7 @@
 }
 
 .tech-stack-dialog-content {
-	backdrop-filter: blur(20px);
+	overflow-y: scroll;
 	background-color: rgba(0, 0, 0);
 	border: solid #27272a 2px;
 	border-radius: 10px;
@@ -59,8 +59,14 @@
 		rgba(50, 50, 93, 0.25) 0px 30px 60px -12px inset,
 		rgba(0, 0, 0, 0.3) 0px 18px 36px -18px inset;
 	position: fixed;
-	top: 6em;
+	top: 2em;
 	left: 1em;
 	width: 95%;
 	height: 90%;
+}
+
+.junit-img {
+	height: 140px;
+	margin-top: 3rem;
+	border-radius: 35px;
 }
