+++ src/pages/AndrewNorik/DevTeamOne.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import AboutMe from '../../components/AboutMeNZ';
-import AboutMeAG from '../../components/AboutMeAG';
-import Carousel from '../../components/Carousel';
-import TechStackDialog from '../..//components/TechStackDialog';
+import AboutMe from '../../components/AboutMeAll/AboutMeNZ';
+import AboutMeAG from '../../components/AboutMeAll/AboutMeAG';
+import Carousel from '../../components/ui/Carousel';
+import TechStackDialog from '../../components/ui/TechStackDialog';
 
 import './styles/index.css';
