+++ src/pages/AndrewNorik/styles/index.css
@@ -1,5 +1,6 @@
 .section-nz {
 	height: 100vh;
+	margin-bottom: 100px;
 }
 
 .wrapper {
