+++ src/pages/TasneemSamiul/DevTeamTwo.tsx
@@ -1,12 +1,12 @@
-import React from 'react'
-import AboutMeSam from '../../components/AboutMeSamTas'
+import React from 'react';
+import AboutMeSam from '../../components/AboutMeAll/AboutMeSamTas';
 
 const DevTeamTwo = () => {
 	return (
 		<>
 			<AboutMeSam />
 		</>
-	)
-}
+	);
+};
 
-export default DevTeamTwo
+export default DevTeamTwo;
